# By Claudio (19) and Claudio Bisegni (2)
# Via Claudio
* commit '862c065f95a6d7581e0a8e7908dbce47e1404948': (21 commits)
  fix cmake compilation
  various fix
  added various fix and improvement:     - ChaosLabel on CCS now check if key are present     - NodeRegister MDS Api now don't print the received data pack     - MDSBatchCommand and MDSBatchExecutor now include the priority on submit command api     - on RegistrationACK for control unit now use priority when submit the two command (init is executed before start)
  fix for deadlock when a command enqueue more then one other command in same sandbox
  try to fix dead loakwhen two command are enqueued in sequence [to test]
  fixed the submition and now the thread scehduler and timeout of the commands are set by default on MDSBatchCommand base class
  ported the init/deinit/start/stop phase within batch command, instead api. This will permit to achive this chage of cu state within other batch command. The batch command needs to be acces the health data to wait until control unit has reaced
  fixed batch system to permit mds command to forward other batch command
  changed the batch command executor to permit the submition of command using submition parameter in method and CDataWrapper only for command data. This will be used in metadata server for submit command within other command
  optimized color representation of the state in CHaosLabel(CCS), correct a bug in the recovery phase of state machine
  fixed the recoverable errors beahviour,
  fix bug that prevent to close monitor on ChaosLabel
  fix compilation issues
  added the api for recover form RecoverableError in (CU, MDS, MDS client)
  fix
  added the visualization of status error in chaos label tooltip
  debuged fatal error during define, init, and start
  fixed the mds on create a complete and wel formend load control unit datapack when use load api
  added MDS api for copy or move  a cu instance, in ccs has been implemented only the copy operation of an instance to another instance of the same unit server. remain to test the error state
  change has been made on control unit to sue new error state fof the state machine. Now when an exception is throwed by framework the cu go in fatal error, all exception thrown by user with run and correlation method(rt and sc cu) will be treated as recoverable error. nned to be tested
  ...

+++ chaos/common/utility/InizializableService.h
@@ -80,7 +80,7 @@ namespace chaos {
 			class InizializableService {
 				boost::msm::back::state_machine< service_state_machine::id_states_machine > state_machine;
 			protected:
-				uint8_t serviceState;
+				unsigned int serviceState;
 				
 				//! Initialize instance
 				virtual void init(void*) throw(chaos::CException);
